+++ src/Pages/Quiz/Quiz.tsx
@@ -2,17 +2,21 @@ import { Box } from "@mui/material"
 import './style.css';
 import { RoundIndicator } from "../../components/RoundIndicator/RoundIndicator";
 import { QuizBoard } from "../../components/QuizBoard/QuizBoard";
-import { useState } from "react";
+import { useGetGameCounter } from "../../hooks/useGetGameCounters";
 
 export const Quiz = () => {
-    const [actualQuestionIndex, setActualQuestionIndex] = useState<number>(0);
-    const [score, setScore] = useState<number>(0);
+    const {
+        actualQuestionIndex,
+        score,
+        setActualQuestionIndex,
+        setScore,
+    } = useGetGameCounter();
+    
     return <Box className="quizPage">
         <RoundIndicator actualQuestionIndex={actualQuestionIndex} score={score}/>
         <QuizBoard 
             actualQuestionIndex={actualQuestionIndex} 
             setActualQuestionIndex={setActualQuestionIndex}
-            score={score}
             setScore={setScore}
         />
     </Box>
